refactor(kernel): introduce x86-generic interrupts interface
