+++ src/kernel/arch/x86/interrupts.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void
+interrupts_init();
